msg/async: dissect AES GCM handlers into dedicated header.
The rationale of this change is to prepare for unit testing.
There is no vital change to the implementation; it just about
shuffling the code.
